Set Markdown SVG default background to white
This is needed as many SVG images have a
transparent background (or no background, only
the strokes) and do no render well in dark mode otherwise.

Cherry-picked from reverted #30, and while we are at it satisfy
stylelint (use `#fff` instead of white), and part of Copilot's
suggested changes (more semantic naming of the class used for svg
images from markdown).

Using a CSS custom property (variable) is a little more involved to do
properly, and will depend on a change to SCSS processing, so is not
included in this commit.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>